proto: Index by model name rather than GVK
Remove completely the notion of Group-version-kind from this package,
as we'll move it where it's actually relevant. This removes the
parsing/constructing which was kind of out of place.

Also allow to return the full list of models. The code in kubernetes
will get all these models, and then index by gvk.