Fix regexp cache race issue
The original implementation allows for read to happen without acquiring lock.
This leads to data race when read and write happen concurrently.

Although tempting, first implementation used sync.Map, which optimize for a map that's used as cache concurrently.
Later changed it back to Mutex for support of go 1.8 as sync.Map is introduced in go 1.9

Signed-off-by: Victor Xian Wang <victor.xianwang@gmail.com>