Add DeepCopy and DeepCopyInto methods to all types
Kubernetes has a project called code-generator[0] which generates
DeepCopy methods for API models.

Cilium[1] is using code-generator on top of models generated by
go-swagger, to allow deep copying of API models. Hovewer, code-generator
assumes that every type which is not a basic type in Golang, needs to
have its own DeepCopy and DeepCopyIn methods. It includes also type
extensions which are built on top of basic types just to add methods.

Currently, if code-generator is executed on go-swagger API models which
use types from strfmt, generated DeepCopy methods fail, because strfmt
types do not implement DeepCopy.

This change adds implementations of DeepCopy and DeepCopyInto which will
allow to use go-generator together with go-swagger and strfmt.

[0] https://github.com/kubernetes/code-generator
[1] https://github.com/cilium/cilium

Signed-off-by: Michal Rostecki <mrostecki@suse.de>