Use google/uuid not pborman/uuid
pborman maintains google's uuid package, and they have a complicated shared history,
but the release notes for pborman/uuid now read

The choice of this package vs the google/uuid package is largely dependent on your
use case. If you want to use UUIDs as keys, used the google/uuid version (a UUID
is an array). Use this package for compatibility, needing to distinguish between
non-set UUID and the zero UUID, prefer byte slices over arrays.

google/uuid is now a dependency of pborman/uuid which is now just a compat wrapper,
so switch to just using google/uuid. Changes are just minor API differences.

Signed-off-by: Justin Cormack <justin.cormack@docker.com>