Fixes issues related to swagger-specific rules with array schemas and items.
Behavior of the swagger validation is the one expected by swagger (i.e. arrays must have items, items must be in arrays).

Default behavior of the json-schema validation is the one expected by jsonschema.

The behavior expected for swagger schemas is enabled by an option to check specifically for arrays and items.

Additional testing is carried out to check for different situations: in example, in response examples, in default values.

* extra linting work, particularly on tests.
* updated deps with other go-openapi repos
* updated golanci-lint config and disable latest round of new, unused linters

* fixes #43  (property named "items", in examples)
* fixes #83  (pure jsonschema validation of array schemas without items specification)
* fixes #108 (property named "items")
* fixes #112 (property named "items")
* fixes go-swagger/go-swagger#1456

Signed-off-by: Frederic BIDON <fredbi@yahoo.com>