Add API Linter for validating lifecycle comment tags
This commit adds an API linter which takes a list of filenames containing
feature gate information. The linter then parses these files and
generates a `featureGatesMap` storing information about the feature gate
name and its corresponding (highest) status and minimum k8s version for
the status.

These values are validated against the values provided in the lifecycle
comment tag.

Note that this linter is very k/k specific and should not be modified
without ensuring that k/k passes any changes to this linter.