Decouple integretion tests from GOPATH
The header file is loaded from the GOPATH
k8s dir by default, which makes it impossible
to develop if you do not have the local
dependencies in your GOPATH.

This commit ensures the header file is
loaded from the project's directory.

It also includes a cleanup of the README
and bumps the timeout to 10s.

Signed-off-by: austin ce <austin.cawley@gmail.com>