Set path for all specific validators when set on SchemaValidator
Currently if we set the path for the SchemaValidator after creation it
does not propagate down to the specific validators embedded within. That
means that subsequent SchemaValidators that are created by the specific
validators will have the wrong path. This updates to set the path for
all specific validators when it is set on their parent SchemaValidator.

Signed-off-by: hasheddan <georgedanielmangum@gmail.com>