Set path for schema validators embedded in schema props validators
Setting the path of the schemaPropsValidator currently does not set the
paths for any of the embedded SchemaValidators. This is primarily an
issue when the parent SchemaValidator sets the path of its embedded
schemaPropsValidator after it is already constructed. In that case, the
path is not propagated to the schemaPropsValidator's embededded
SchemaValidators (which will now in turn set the paths for all of their
embedded validators).

Signed-off-by: hasheddan <georgedanielmangum@gmail.com>