Do not import github.com/google/go-cmp/cmp outside tests.
Go-cmp relies heavily on reflection in ways that casue the golang
compiler to disable the dead code elimination. Even declaring a
variable of type cmp.Option disables the DCE.

fuzz.go is only used in tests, but is placed in the package itself.
This disables the DCE in every user of k8s.io/client-go. Move it to
fuzz_test.go instead.